server/requirements.txt
Flask==2.3.2
opencv_python_headless==4.8.0.74
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
fastapi==0.100.1
filelock==3.12.2
gdown==4.7.1
h11==0.14.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
PySocks==1.7.1
python-multipart==0.0.6
pywatchman==1.4.1
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.12
tifffile==2023.7.18
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
